dashApp_secopi/documents/requirements.txt
asn1crypto==1.2.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
boto==2.49.0
boto3==1.10.6
botocore==1.13.6
bz2file==0.98
catalogue==0.0.8
certifi==2019.11.28
cffi==1.13.1
chardet==3.0.4
chart-studio==1.0.0
Click==7.0
colorlover==0.3.0
cryptography==2.7
cufflinks==0.13.0
cycler==0.10.0
cymem==2.0.3
dash==1.4.1
dash-bootstrap-components==0.7.2
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.1
gensim==3.8.0
geographiclib==1.50
geopy==1.20.0
gunicorn==20.0.0
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
jsonschema==3.1.1
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
llvmlite==0.30.0
lorem==0.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==7.2.0
murmurhash==1.0.0
nbconvert==5.6.1
nbformat==4.4.0
netifaces==0.10.9
nltk==3.4.5
notebook==6.0.1
numba==0.46.0
numpy==1.17.3
olefile==0.46
packaging==19.2
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plac==0.9.6
plotly==4.2.1
plotly-express==0.4.1
ply==3.11
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyarrow==0.13.0
pycparser==2.19
Pygments==2.4.2
pymapd==0.18.0
pyOpenSSL==19.0.0
pyparsing==2.4.5
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyrsistent==0.15.5
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
rbc-project==0.1.1.dev0
requests==2.22.0
retrying==1.3.3
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
Shapely==1.6.4.post2
simple-salesforce==0.74.2
sip==4.19.18
six==1.12.0
smart-open==1.9.0
spacy==2.2.3
SQLAlchemy==1.3.10
srsly==0.2.0
statsmodels==0.10.1
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
thinc==7.3.0
thrift==0.11.0
thriftpy2==0.4.9
tornado==6.0.3
tqdm==4.40.0
traitlets==4.3.3
urllib3==1.25.6
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordcloud==1.6.0
zipp==0.6.0
